# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Blossom Buddy API",
        "description": "API pour la gestion des plantes",
        "contact": {
            "email": "admin@blossombuddy.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local"
        }
    ],
    "paths": {
...
        "/plant": {
            "get": {
                "tags": [
                    "Plant"
                ],
                "summary": "Liste toutes les plantes",
                "operationId": "d5241537f6505ae4a717f0c92bbb40fb",
                "responses": {
                    "200": {
                        "description": "Liste des plantes"
                    }
                }
            },
            "post": {
                "tags": [
                    "Plant"
                ],
                "summary": "Ajoute une nouvelle plante",
                "operationId": "8309c05e5772465302dd924140b3f526",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "common_name": {
                                        "type": "string",
                                        "example": "Test"
                                    },
                                    "ville": {
                                        "type": "string",
                                        "example": "Paris"
                                    },
                                    "watering": {
                                        "type": "string",
                                        "example": "medium"
                                    },
                                    "watering_general_benchmark": {
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Plante créée et météo du lieu",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "plant": {
                                            "type": "object"
                                        },
                                        "weather": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/plant/{name}": {
            "get": {
                "tags": [
                    "Plant"
                ],
                "summary": "Affiche une plante par son nom",
                "operationId": "80c395237d1c7867d3775bc0bdbb3736",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails de la plante"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Plant"
                ],
                "summary": "Supprime une plante par son nom",
                "operationId": "f4a7fece79a12719e88ec0c9f7d1faa7",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Plante supprimée"
                    }
                }
            }
        },
        "/plant/watering-schedule": {
            "post": {
                "tags": [
                    "Plant"
                ],
                "summary": "Calcule le temps avant le prochain arrosage pour une plante",
                "operationId": "44fbd74eea1e4a5efdd8e309f68cfcad",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "plant_common_name": {
                                        "type": "string",
                                        "example": "Rose"
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "Paris"
                                    },
                                    "last_watered_at": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2025-10-06T10:00:00Z"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Calcul du prochain arrosage",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user_plant": {
                                            "type": "object"
                                        },
                                        "watering_schedule": {
                                            "type": "object"
                                        },
                                        "weather_data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/plant/water": {
            "post": {
                "tags": [
                    "Plants"
                ],
                "summary": "Enregistrer un arrosage de plante",
                "description": "Met à jour la date du dernier arrosage pour une plante de l'utilisateur",
                "operationId": "waterPlant",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "plant_id",
                                    "city"
                                ],
                                "properties": {
                                    "plant_id": {
                                        "description": "ID de la plante",
                                        "type": "integer"
                                    },
                                    "city": {
                                        "description": "Ville où se trouve la plante",
                                        "type": "string"
                                    },
                                    "watered_at": {
                                        "description": "Date et heure d'arrosage (optionnel, par défaut maintenant)",
                                        "type": "string",
                                        "format": "date-time"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Arrosage enregistré avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "user_plant": {
                                            "type": "object"
                                        },
                                        "next_watering_date": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Relation plante-utilisateur non trouvée"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/user/plants-with-schedule": {
            "get": {
                "tags": [
                    "Plants"
                ],
                "summary": "Récupérer les plantes de l'utilisateur avec leur programme d'arrosage",
                "description": "Retourne toutes les plantes de l'utilisateur avec leur prochain arrosage calculé",
                "operationId": "getUserPlantsWithSchedule",
                "parameters": [
                    {
                        "name": "city",
                        "in": "query",
                        "description": "Filtrer par ville (optionnel)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des plantes avec programme d'arrosage",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "plant": {
                                                "type": "object"
                                            },
                                            "watering_schedule": {
                                                "type": "object"
                                            },
                                            "weather_data": {
                                                "type": "object"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
}